Move server-side storage code to a more reasonable spot
